requirements.txt
google-cloud-bigquery
requests
functions-framework
python-dotenv